+++ lib/kubernetes-deploy/kubernetes_resource/pod.rb
@@ -9,6 +9,8 @@ module KubernetesDeploy
       Preempting
     )
 
+    attr_accessor :stream_logs
+
     def initialize(namespace:, context:, definition:, logger:,
       statsd_tags: nil, parent: nil, deploy_started_at: nil, stream_logs: false)
       @parent = parent